byte

+++ src/com/sort/hong/BitMap.java
@@ -6,10 +6,7 @@ public class BitMap {
 	public static void main(String[] args) {
 		int [] before = {1,3,4,3,6,7,34,76,23,676,446,7,454,343,67,99,3,2,88,6567,90};
 		bitSort(before);
-		
 	}
-	
-	
 	static void bitSort(int[] a) {
 		int min = a[0];
 		int max = a[0];
@@ -19,10 +16,9 @@ public class BitMap {
 			
 		}
 		int len = max-min +1;
-		
-		int [] after = new int[len];
+		byte [] after = new byte[len];
 		for(int b:a) {
-			after[b-min] = after[b-min]+1;
+			after[b-min] = (byte) (after[b-min]+1);
 		}
 		
 		
@@ -30,7 +26,7 @@ public class BitMap {
 				while(after[i]>0) {
 					System.out.print(i+min);
 					System.out.print(",");
-					after[i] = after[i]-1;
+					after[i] = (byte) (after[i]-1);
 				}
 		}
